backend/requirements.txt
fastapi
uvicorn
pyyaml

